requirements.txt
absl-py==1.4.0
ar-corrector==1.1.6
arabic-reshaper==3.0.0
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
bidict==0.23.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
djangorestframework==3.15.1
dm-tree==0.1.8
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.35.2
executing==1.2.0
filelock==3.13.1
Flask==2.3.2
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.41.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.56.0
h11==0.14.0
h5py==3.9.0
idna==3.4
imutils==0.5.4
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.25
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==3.0.4
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.2.0
namex==0.0.7
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scipy==1.11.1
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
stack-data==0.6.2
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.25.1
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
wsproto==1.2.0
